redux로 리팩토링 완료
