+++ src/components/LetterDetailWrapper.jsx
@@ -1,21 +1,23 @@
-import { deleteMemberLetter, updateMemberLetter } from "modules/memberLetters";
-import { useRef, useState } from "react";
-import { useDispatch } from "react-redux";
-import { useNavigate } from "react-router-dom";
+import { alter } from "lib/alter";
+import {
+  deleteMemberLetter,
+  selectMemberLetterList,
+  updateMemberLetter,
+} from "modules/memberLetters";
+import { useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { useNavigate, useParams } from "react-router-dom";
 import Button from "./Button";
 import LetterDetail from "./LetterDetail";
 
-export default function LetterDetailWrapper({
-  writedTo,
-  content,
-  nickname,
-  createdAt,
-  avatar,
-  id,
-}) {
+export default function LetterDetailWrapper() {
+  const { member, id } = useParams();
+  const memberLetterList = useSelector(selectMemberLetterList);
+  const letter = memberLetterList[member].find(findLetterById(id));
+  const ifEmptyLetterThan = alter(() => !letter);
+
   const [isUpdate, setIsUpdate] = useState(false);
-  const [contentValue, setContentValue] = useState(content);
-  const contentRef = useRef(null);
+  const [contentValue, setContentValue] = useState(letter.content);
   const navigate = useNavigate();
 
   const dispatch = useDispatch();
@@ -23,7 +25,7 @@ export default function LetterDetailWrapper({
   const handleUpdateButton = () => setIsUpdate(true);
   const handleUpdateDoneButton = () => {
     setIsUpdate(false);
-    if (contentValue === content) {
+    if (contentValue === letter.content) {
       alert("아무런 수정 사항이 없습니다.");
       return;
     }
@@ -47,20 +49,24 @@ export default function LetterDetailWrapper({
     <Button handleClickEvent={handleUpdateDoneButton}>수정 완료</Button>
   );
 
-  return (
+  return ifEmptyLetterThan(
+    <EmptyLetterDetail />,
     <LetterDetail
+      {...letter}
       RemoveButton={RemoveButton}
       UpdateButton={UpdateButton}
       UpdateDoneButton={UpdateDoneButton}
-      avatar={avatar}
-      content={content}
-      contentRef={contentRef}
-      contentValue={contentValue}
-      createdAt={createdAt}
       handleChangeContent={handleChangeContent}
       isUpdate={isUpdate}
-      nickname={nickname}
-      writedTo={writedTo}
-    />
+      contentValue={contentValue}
+    />,
   );
 }
+const EmptyLetterDetail = () => {
+  return (
+    <div>
+      <h2>You've taken the wrong path.</h2>
+    </div>
+  );
+};
+const findLetterById = (id) => (letter) => letter.id === id;
